requirements.txt
langchain==0.1.16
langchain_openai==0.1.3
python-dotenv==1.0.1
streamlit==1.33.0
